+++ frontend/src/components/google-maps/marker/RegionCountMarkersContainer/components/RegionCountMarker.style.ts
